Carry header chain verification to da verifier
